Move UIViewController snapshot helpers into helper class
